Handle missing variables + get_lon_lat_time
Add missing function get_lon_lat_time required in the Main_workshop (load the lon, lat and time from specified floats).

Correct the plot functions and get_multi_profile_mean to be able to deal with selections of floats that does not all show the wanted variables.

Correction of the bug in the plot_profile function to actually plot all profiles of selected floats (and not only one float data).

Modify the depth_interp function so it can skip profiles with not enough PRES values